+++ redisgraph_bulk_loader/bulk_insert.py
@@ -5,7 +5,7 @@ import click
 from timeit import default_timer as timer
 
 sys.path.append(os.path.dirname(__file__))
-from config import Config
+from config import Config, Config_Set
 from query_buffer import QueryBuffer
 from label import Label
 from relation_type import RelationType
@@ -38,25 +38,6 @@ def process_entities(entities):
         # Add binary data to list and update all counts
         entity.query_buffer.redis_token_count += len(entity.binary_entities)
         entity.query_buffer.buffer_size += added_size
-
-
-def Config_Set(max_token_count, max_buffer_size, max_token_size, enforce_schema, skip_invalid_nodes, skip_invalid_edges, separator, quoting):
-    # Maximum number of tokens per query
-    # 1024 * 1024 is the hard-coded Redis maximum. We'll set a slightly lower limit so
-    # that we can safely ignore tokens that aren't binary strings
-    # ("GRAPH.BULK", "BEGIN", graph name, counts)
-    Config.max_token_count = min(max_token_count, 1024 * 1023)
-    # Maximum size in bytes per query
-    Config.max_buffer_size = max_buffer_size * 1000000
-    # Maximum size in bytes per token
-    # 512 megabytes is a hard-coded Redis maximum
-    Config.max_token_size = min(max_token_size * 1000000, 512 * 1000000)
-
-    Config.enforce_schema = enforce_schema
-    Config.skip_invalid_nodes = skip_invalid_nodes
-    Config.skip_invalid_edges = skip_invalid_edges
-    Config.separator = separator
-    Config.quoting = quoting
 
 
 # Command-line arguments
